pyproject.toml
[tool.poetry]
name = "extremeflash"
version = "0.1.0"
description = ""
authors = ["Grische <2787581+grische@users.noreply.github.com>"]
license = "GPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6"
pyserial = "^3.5"
tftpy = "^0.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
